refactor: alinear configuración y validaciones de la aplicación
- Ajustar configuraciones de entidades para alinearlas con PostgreSQL.
- Corregir claves primarias, foráneas, índices únicos y valores por defecto.
- Alinear las entidades de dominio con la definición actual de la base de datos.
- Eliminar el validador independiente de organizaciones.
- Centralizar la consulta de organizaciones mediante IOrganizationRepository.
- Actualizar los handlers de organizaciones, fechas de negocio y módulos.
- Garantizar la validación de existencia de la organización desde los casos de uso.
- Ajustar la relación jerárquica de módulos mediante clave foránea compuesta.
- Eliminar la consulta independiente de módulo por identificador.
- Ajustar la consulta de módulos activos y su construcción jerárquica.
- Incorporar ConfigurationException para errores de configuración del sistema.
- Diferenciar recursos inexistentes de configuraciones obligatorias no configuradas.
- Actualizar el middleware para manejar excepciones de configuración.
- Mejorar la verificación de conexión a la base de datos durante el startup.

+++ src/Prestamax.Domain/Configuration/Settings/Setting.cs
@@ -5,9 +5,9 @@
 /// </summary>
 public sealed class Setting
 {
-    public long IdSetting { get; private set; }
+    public int IdSetting { get; private set; }
 
-    public long IdOrganization { get; private set; }
+    public int IdOrganization { get; private set; }
 
     public string SettingKey { get; private set; } = string.Empty;
